fix #944: use long instead of int for comment ids

+++ src/org/wordpress/android/ui/comments/CommentsListFragment.java
@@ -250,7 +250,7 @@ public class CommentsListFragment extends Fragment {
         CommentActions.moderateComments(WordPress.getCurrentLocalTableBlogId(), selectedComments, CommentStatus.TRASH, listener);
     }
 
-    protected void setHighlightedCommentId(int commentId) {
+    protected void setHighlightedCommentId(long commentId) {
         getCommentAdapter().setHighlightedCommentId(commentId);
     }
